requirements.txt
pytest
flake8
black
pillow
tqdm
